+++ src/lib/exports.js
@@ -1,0 +1,26 @@
+export {
+  getAuth,
+  onAuthStateChanged,
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+  signInWithPopup,
+  GoogleAuthProvider,
+  sendPasswordResetEmail,
+  updateProfile,
+  signOut,
+  } from 'https://www.gstatic.com/firebasejs/9.10.0/firebase-auth.js'; //eslint-disable-line
+
+export { initializeApp } from 'https://www.gstatic.com/firebasejs/9.10.0/firebase-app.js';//eslint-disable-line
+
+export {
+  getFirestore,
+  collection,
+  addDoc,
+  getDoc,
+  doc,
+  updateDoc,
+  getDocs,
+  deleteDoc,
+} from 'https://www.gstatic.com/firebasejs/9.10.0/firebase-firestore.js';//eslint-disable-line
+
+// getDocs from firebase-firestore.js
