+++ src/pages/login/main.js
@@ -1,7 +1,7 @@
 import {
   loginWithEmailAndPassword,
   loginWithGoogle,
-} from '../../lib/index.js';
+} from '../../lib/firebase.js';
 
 import {
   handleFirebaseErrors,
