+++ src/pages/resetPassword/main.js
@@ -1,6 +1,6 @@
 import {
   resetPassword,
-} from '../../lib/index.js';
+} from '../../lib/firebase.js';
 
 import {
   handleFirebaseErrors,
